Rename AssessmentReportInstanceList
