+++ src/DetailsView/components/tab-stops/tab-stops-instance-section-props-factory.tsx
@@ -9,7 +9,7 @@ import { TabStopsRequirementsWithInstancesProps } from 'DetailsView/tab-stops-re
 import * as styles from 'DetailsView/tab-stops-requirements-with-instances.scss';
 import * as React from 'react';
 import { InstanceReportModel } from 'reports/assessment-report-model';
-import { AssessmentReportInstanceList } from 'reports/components/assessment-report-instance-list';
+import { ReportInstanceList } from 'reports/components/report-instance-list';
 import { TabStopRequirementId } from 'types/tab-stop-requirement-info';
 
 export type TabStopsInstanceSectionPropsFactoryDeps = TabStopsFailedInstanceSectionDeps;
@@ -151,7 +151,7 @@ export const ReportTabStopsInstanceSectionPropsFactory: TabStopsInstanceSectionP
                     />
                 ),
                 content: (
-                    <AssessmentReportInstanceList
+                    <ReportInstanceList
                         key={`${result.id}-requirement-group`}
                         requirementId={result.id}
                         instances={instances}
